requirements.txt
openai>=1.0.0
streamlit
python-dotenv
fastapi
uvicorn[standard] 